Added some code documentation, fixed some bugs to do with the database, made so the bot token is read from database.
