+++ cmds/reddit.go
@@ -1,8 +1,8 @@
 package cmds
 
 import (
-	"../core"
-	"../rest"
+	"github.com/Member1221/plutobot-go/core"
+	"github.com/Member1221/plutobot-go/rest"
 )
 
 type SubredditPosts struct {
@@ -67,3 +67,84 @@ var Connection = prest.New()
 func RedditCommand(a core.CommandArgs, v []string) bool {
 	return true
 }
+
+
+// TODO: Reimplement this.
+/*
+func RedditUpdate(subreddit string) bool {
+	req, err := http.NewRequest("GET", "https://www.reddit.com/r/"+subreddit+"/new.json", nil)
+	if err != nil {
+		fmt.Println(err)
+		return false
+	}
+	req.Header.Set("User-Agent", "UGx1dG9Cb3Q=")
+
+	// Handle the request
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		fmt.Println(err)
+		return false
+	}
+	defer resp.Body.Close()
+	if resp.StatusCode != http.StatusOK {
+		fmt.Println(resp.Status)
+		return false
+	}
+
+	str, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		fmt.Println(err)
+		return false
+	}
+
+	var posts cmds.SubredditPosts
+
+	err = json.Unmarshal(str, &posts)
+	if err != nil {
+
+		fmt.Println(err.Error() + "\n: " + string(str))
+		return false
+	}
+	ASYNC_MUTEX.Lock()
+	if LAST_REDDIT[subreddit] == posts.Data.Children[0].Data.ID {
+		return true
+	}
+	ASYNC_MUTEX.Unlock()
+
+	var thumbnail *discordgo.MessageEmbedThumbnail = nil
+
+	if posts.Data.Children[0].Data.Thumbnail != "self" {
+		thumbnail = &discordgo.MessageEmbedThumbnail{posts.Data.Children[0].Data.Thumbnail, posts.Data.Children[0].Data.Thumbnail, 128, 128}
+	}
+	text := posts.Data.Children[0].Data.Selftext
+	if len(text) > 250 {
+		text = text[:247] + "..."
+	}
+	embed := discordgo.MessageEmbed{
+		"https://reddit.com" + posts.Data.Children[0].Data.Permalink,
+		"rich",
+		posts.Data.Children[0].Data.Title,
+		text,
+		"",
+		0xFF00FF,
+		nil,
+		nil,
+		thumbnail,
+		nil,
+		nil,
+		nil,
+		[]*discordgo.MessageEmbedField{},
+	}
+	params := discordgo.WebhookParams{"", "PlutoBot->Reddit", "", false, "", []*discordgo.MessageEmbed{&embed}}
+
+	err = dg.WebhookExecute(WEBHOOK_REDDIT.ID, WEBHOOK_REDDIT.Token, true, &params)
+	if err != nil {
+		fmt.Println(err.Error())
+		return false
+	}
+	ASYNC_MUTEX.Lock()
+	LAST_REDDIT[subreddit] = posts.Data.Children[0].Data.ID
+	ASYNC_MUTEX.Unlock()
+	return true
+}
+*/
